Refactor RollerIntakeBase to SimpleMotorBase
